+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [18.2.0] - 2018-11-29
+### Changed
+- Components:
+    - `sign-up-form` - Distinguish between alrteady registered and invalid (e.g. blacklisted) emails
+- Models:
+    - `user-registration` - added invalid email validation and `addInvalidEmail` method
+- Routes:
+    - `register` - let CAS redirect to ORCID
+
 ## [18.1.2] - 2018-11-05
 - Engines:
     - `registries/discover` - reset to first page on user search input
@@ -18,10 +27,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Components:
     - `panel` - a general-use abstraction of bootstrap panels
+    - `validated-model-form` - wraps common logic for forms made with `validated-input/*`
 - Tests:
     - `panel` component integration test
 - Handbook:
     - `panel` component
+- Mirage:
+    - `node` POST view to add currentUser as contributor
+    - `regions` fixtures
+    - `wb` view to move files from user or node to a node
+- Routes:
+    - `settings.applications` - list of developer apps
+    - `settings.applications.edit`
+    - `settings.applications.create`
+
 ### Changed
 - Components:
     - `loading-indicator` - added inline option
@@ -31,6 +50,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `loading-indicator` - added tests for inline
 - Handbook:
     - `loading-indicator` - added examples for inline
+- Mirage:
+    - `root` factory now adds all feature flags, not just route flags
+    - `user` factory has 'withFiles' trait so non-current users can have files easily
+    - `user` serializer has default_region relationship (hardcoded to us)
+
 ### Removed
 - Flags:
     - `ember_project_forks_page` - `guid-node.forks` and `guid-registration.forks` now always on
@@ -42,6 +66,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `registration-schema` (including related adapter & serializer)
     - `token`
     - `scope`
+    - `developer-app`
 - Components:
     - `search-help-modal` - you know, the search help modal but as it's own component
     - `draft-registration-card` - summary card for draft registrations
@@ -54,7 +79,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `delete-button` - configurable delete button, including a confirmation modal and scientist name
     - `tags-widget` - you know, for tags
 - Routes:
-    - `guid-node/registrations` - registrations tab
+    - `guid-node.registrations` - registrations tab
     - `settings` - includes the settings side nav
     - `settings/tokens` - list of personal access tokens
     - `settings/tokens/edit`